+++ src/org/ojalgo/matrix/decomposition/EigenvalueDecomposition.java
@@ -21,6 +21,7 @@
  */
 package org.ojalgo.matrix.decomposition;
 
+import org.ojalgo.ProgrammingError;
 import org.ojalgo.array.Array1D;
 import org.ojalgo.matrix.decomposition.function.ExchangeColumns;
 import org.ojalgo.matrix.store.MatrixStore;
@@ -60,6 +61,7 @@ abstract class EigenvalueDecomposition<N extends Number> extends GenericDecompos
 
     private MatrixStore<N> myD = null;
     private Array1D<ComplexNumber> myEigenvalues = null;
+    private int mySquareDim = 0;
     private MatrixStore<N> myV = null;
     private boolean myValuesOnly = false;
 
@@ -121,6 +123,8 @@ abstract class EigenvalueDecomposition<N extends Number> extends GenericDecompos
         myV = null;
 
         myValuesOnly = false;
+
+        mySquareDim = 0;
     }
 
     private final boolean compute(final Access2D.Collectable<N, ? super PhysicalStore<N>> matrix, final boolean hermitian, final boolean valuesOnly) {
@@ -130,6 +134,12 @@ abstract class EigenvalueDecomposition<N extends Number> extends GenericDecompos
         myValuesOnly = valuesOnly;
 
         boolean retVal = false;
+
+        long countRows = matrix.countRows();
+        if (matrix.countColumns() != countRows) {
+            ProgrammingError.throwIfNotSquare(matrix);
+        }
+        mySquareDim = Math.toIntExact(countRows);
 
         try {
 
@@ -154,6 +164,26 @@ abstract class EigenvalueDecomposition<N extends Number> extends GenericDecompos
     protected abstract boolean doGeneral(final Collectable<N, ? super PhysicalStore<N>> matrix, final boolean eigenvaluesOnly);
 
     protected abstract boolean doHermitian(final Collectable<N, ? super PhysicalStore<N>> matrix, final boolean eigenvaluesOnly);
+
+    @Override
+    protected int getColDim() {
+        return mySquareDim;
+    }
+
+    @Override
+    protected int getMaxDim() {
+        return mySquareDim;
+    }
+
+    @Override
+    protected int getMinDim() {
+        return mySquareDim;
+    }
+
+    @Override
+    protected int getRowDim() {
+        return mySquareDim;
+    }
 
     protected abstract MatrixStore<N> makeD();