+++ src/org/ojalgo/matrix/decomposition/LUDecomposition.java
@@ -85,6 +85,20 @@ abstract class LUDecomposition<N extends Number> extends InPlaceDecomposition<N>
         return this.getDeterminant();
     }
 
+    public int countSignificant(final double threshold) {
+
+        DecompositionStore<N> internal = this.getInPlace();
+
+        int significant = 0;
+        for (int ij = 0, limit = this.getMinDim(); ij < limit; ij++) {
+            if (Math.abs(internal.doubleValue(ij, ij)) > threshold) {
+                significant++;
+            }
+        }
+
+        return significant;
+    }
+
     public boolean decompose(final Access2D.Collectable<N, ? super PhysicalStore<N>> matrix) {
         return this.doDecompose(matrix, true);
     }
@@ -136,21 +150,12 @@ abstract class LUDecomposition<N extends Number> extends InPlaceDecomposition<N>
         return myPivot.getOrder();
     }
 
-    public int getRank() {
+    public double getRankThreshold() {
 
-        int retVal = 0;
+        N largest = this.getInPlace().aggregateDiagonal(Aggregator.LARGEST);
+        double epsilon = this.getDimensionalEpsilon();
 
-        final DecompositionStore<N> internalStore = this.getInPlace();
-
-        final double largestValue = internalStore.aggregateAll(Aggregator.LARGEST).doubleValue();
-
-        for (int ij = 0, limit = this.getMinDim(); ij < limit; ij++) {
-            if (!internalStore.isSmall(ij, ij, largestValue)) {
-                retVal++;
-            }
-        }
-
-        return retVal;
+        return epsilon * Math.max(MACHINE_SMALLEST, largest.doubleValue());
     }
 
     public final MatrixStore<N> getSolution(final Collectable<N, ? super PhysicalStore<N>> rhs) {
@@ -213,10 +218,6 @@ abstract class LUDecomposition<N extends Number> extends InPlaceDecomposition<N>
         } else {
             throw RecoverableCondition.newMatrixNotInvertible();
         }
-    }
-
-    public boolean isFullRank() {
-        return this.isSolvable();
     }
 
     public boolean isPivoted() {
@@ -306,7 +307,8 @@ abstract class LUDecomposition<N extends Number> extends InPlaceDecomposition<N>
 
     @Override
     protected boolean checkSolvability() {
-        return (this.getRowDim() == this.getColDim()) && (this.getRank() == this.getColDim());
+        double threshold = Math.min(this.getRankThreshold(), MACHINE_EPSILON);
+        return (this.getRowDim() == this.getColDim()) && (this.getColDim() == this.countSignificant(threshold));
     }
 
     int[] getReducedPivots() {