+++ src/org/ojalgo/matrix/store/MatrixPipeline.java
@@ -66,6 +66,34 @@ abstract class MatrixPipeline<N extends Number> implements ElementsSupplier<N> {
         }
     }
 
+    static final class ColumnsModifier<N extends Number> extends MatrixPipeline<N> {
+
+        private final BinaryFunction<N> myFunction;
+        private final Access1D<N> myRightArgumnts;
+
+        ColumnsModifier(final ElementsSupplier<N> base, final BinaryFunction<N> modifier, final Access1D<N> right) {
+            super(base);
+            myFunction = modifier;
+            myRightArgumnts = right;
+        }
+
+        @Override
+        public void supplyTo(final TransformableRegion<N> receiver) {
+
+            this.getContext().supplyTo(receiver);
+
+            UnaryFunction<N> modifier;
+
+            final long limit = Math.min(receiver.countColumns(), myRightArgumnts.count());
+            for (long j = 0; j < limit; j++) {
+                modifier = myFunction.second(myRightArgumnts.get(j));
+                receiver.modifyColumn(j, modifier);
+            }
+
+        }
+
+    }
+
     static final class ColumnsReducer<N extends Number> extends MatrixPipeline<N> {
 
         private final Aggregator myAggregator;
@@ -115,6 +143,34 @@ abstract class MatrixPipeline<N extends Number> implements ElementsSupplier<N> {
         @Override
         public void supplyTo(final TransformableRegion<N> receiver) {
             receiver.fillByMultiplying(myLeft, myRight);
+        }
+
+    }
+
+    static final class RowsModifier<N extends Number> extends MatrixPipeline<N> {
+
+        private final BinaryFunction<N> myFunction;
+        private final Access1D<N> myRightArgumnts;
+
+        RowsModifier(final ElementsSupplier<N> base, final BinaryFunction<N> modifier, final Access1D<N> right) {
+            super(base);
+            myFunction = modifier;
+            myRightArgumnts = right;
+        }
+
+        @Override
+        public void supplyTo(final TransformableRegion<N> receiver) {
+
+            this.getContext().supplyTo(receiver);
+
+            UnaryFunction<N> modifier;
+
+            final long limit = Math.min(receiver.countRows(), myRightArgumnts.count());
+            for (long i = 0; i < limit; i++) {
+                modifier = myFunction.second(myRightArgumnts.get(i));
+                receiver.modifyRow(i, modifier);
+            }
+
         }
 
     }