+++ CHANGELOG.md
@@ -4,6 +4,14 @@
 
 > Corresponds to changes in the `develop` branch since the last release
 
+## org.ojalgo.data
+
+* Renamed DataPreprocessors to DataProcessors, and added methods to create a covariance matrix from an SVD.
+
+## org.ojalgo.function
+
+* Added a couple more utilities to MissingMath
+
 ## org.ojalgo.matrix
 
 * New LU decomposition implementation that is faster for small matrices.
@@ -11,6 +19,8 @@
 * Renamed checkAndCompute to checkAndDecompose in the MatrixDecomposition.Hermitian interface
 * Extended the MatrixStore.LogicalBuilder rows() and columns() methods to allow for negative indexes that refer to all zero rows/columns.
 * New more efficient implementation of MatrixStore.LogicalBuilder.diagonal() data structure.
+* The MatrixDecomposition interface now extends Structure2D so that you can get the size/shape of the original matrix.
+* Additions to the MatrixDecomposition.RankRevealing interface. It is now possible estimate the rank using a custom threshold. Also reworked all code related estimating rank (in all implementations of that interface).
 
 ## org.ojalgo.netio
 
@@ -18,7 +28,12 @@
 
 ## org.ojalgo.optimisation
 
+* The optimisation model parameter scaling functionality has been tweaked to be more in line with the improved "rank revealing" of matrices.
 * Minor improvements to ConvexSolver
+
+## org.ojalgo.structure
+
+* The Stream2D interface now has methods operateOnColumns(...) and operateOnRows(...)
 
 # v47.1.2: 2019-04-23